backend/queue-service/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pika==1.3.2
flask_cors==5.0.0
